refactoring record.go, file.go
